requirements.txt
torch==1.7.0
torchvision==0.8.0
torchaudio==0.7.0
open3d-python=0.6.0.0
tensorboardX
torchsummary
tqdm
matplotlib
opencv-python==4.3.0.38